fix: created type

+++ async-openai/src/types/image.rs
@@ -129,7 +129,7 @@ pub enum Image {
 
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
 pub struct ImagesResponse {
-    pub created: u32,
+    pub created: u64,
     pub data: Vec<std::sync::Arc<Image>>,
 }
